// packages/database/prisma-generated/zod/inputTypeSchemas/RegistryUserJoinCountOrderByAggregateInputSchema.ts
import type { Prisma } from '../../client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const RegistryUserJoinCountOrderByAggregateInputSchema: z.ZodType<Prisma.RegistryUserJoinCountOrderByAggregateInput> = z.object({
  id: z.lazy(() => SortOrderSchema).optional(),
  createdAt: z.lazy(() => SortOrderSchema).optional(),
  updatedAt: z.lazy(() => SortOrderSchema).optional(),
  role: z.lazy(() => SortOrderSchema).optional(),
  userId: z.lazy(() => SortOrderSchema).optional(),
  registryId: z.lazy(() => SortOrderSchema).optional()
}).strict();

export default RegistryUserJoinCountOrderByAggregateInputSchema;